Updated files
I looked over the files and corrected some of the errors- I also made
the NXT write the speed values of each motor to the screen- motor A is
line one, motor B is line two, ect for troubleshooting purposes.

Don't bother with the Tethered Test.vi- I'm just seeing if I can test
the program without the Samantha.